task evaluation to own file
