+++ src/state/quests/questSaga.ts
@@ -1,13 +1,12 @@
 import * as moment from 'moment';
 import { delay } from 'redux-saga';
 import { call, put, select, takeEvery } from 'redux-saga/effects';
-import { Quest, QuestTask } from '../AppState';
+import { Quest } from '../AppState';
 import { QuestConfig as Config } from '../Config';
-import { getCapital } from '../depot/depotSelector';
 import { calculateNextStockValues } from '../stockMarket/stockMarketActions';
-import { getOwnedStocksAmount } from '../stockMarket/stockSelector';
 import { addQuests, LOAD_QUESTS, RECALCULATE_QUESTS, recalculateQuests, updateQuest } from './questActions';
 import { getActiveQuests } from './questSelectors';
+import { getTaskProgress } from './taskProgressEvaluation';
 
 const questJson = require('./quests.json');
 
@@ -28,27 +27,6 @@ function* loadInitialQuests() {
 
     yield put(addQuests(quests));
     yield put(recalculateQuests());
-}
-
-function* getTaskProgress(task: QuestTask) {
-    switch (task.questType) {
-        case 'moneyPossession':
-            return yield call(getMoneyPossessionTaskProgress, task);
-        case 'StockTotalPossession':
-            return yield call(getStockTotalPossessionTaskProgress, task);
-        default:
-            return 0;
-    }
-}
-
-function* getMoneyPossessionTaskProgress(task: QuestTask) {
-    const money = yield select(getCapital);
-    return money * 100 / task.amount;
-}
-
-function* getStockTotalPossessionTaskProgress(task: QuestTask) {
-    const ownedStockAmount = yield select(getOwnedStocksAmount);
-    return ownedStockAmount * 100 / task.amount;
 }
 
 function* recalculateAllQuests() {
